Finish renamings
