+++ src/admin/project/new/Step1.js
@@ -2,7 +2,7 @@ import React, { useState } from 'react'
 import NewProjectLayout from './NewProjectLayout'
 import { object, string } from 'yup'
 import { Form, Formik } from 'formik'
-import OFFormControlInputFormiked from '../../baseComponents/form/formcontrol2/OFFormControlInputFormiked'
+import OFFormControlInputFormiked from '../../baseComponents/form/formControl/OFFormControlInputFormiked'
 import OFButton from '../../baseComponents/button/OFButton'
 import Box from '@material-ui/core/Box'
 import { useTranslation } from 'react-i18next'
