+++ src/admin/project/settings/event/ProjectSettingsForm.js
@@ -2,15 +2,15 @@ import React, { useState } from 'react'
 import { object, string, boolean } from 'yup'
 import { Typography } from '@material-ui/core'
 import { Field, Form, Formik } from 'formik'
-import OFFormControl from '../../../baseComponents/form/formcontrol2/OFFormControl'
+import OFFormControl from '../../../baseComponents/form/formControl/OFFormControl'
 import Grid from '@material-ui/core/Grid'
 import OFButton from '../../../baseComponents/button/OFButton'
 import makeStyles from '@material-ui/core/styles/makeStyles'
 import { useDispatch } from 'react-redux'
 import { editProject } from '../../core/projectActions'
 import ChipColorsEditor from './ChipColorsEditor'
-import OFFormControlInputFormiked from '../../../baseComponents/form/formcontrol2/OFFormControlInputFormiked'
-import OFDateTimePicker from '../../../baseComponents/form/datetimepicker2/OFDateTimePicker'
+import OFFormControlInputFormiked from '../../../baseComponents/form/formControl/OFFormControlInputFormiked'
+import OFDateTimePicker from '../../../baseComponents/form/dateTimePicker/OFDateTimePicker'
 import FormControlLabel from '@material-ui/core/FormControlLabel'
 import { OFSwitch } from '../../../baseComponents/form/switch/OFSwitch'
 import Collapse from '@material-ui/core/Collapse'
