+++ src/admin/project/setupTypeForms/SetupHoverboardv2Form.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { object, string } from 'yup'
 import { Form, Formik } from 'formik'
-import OFFormControlInputFormiked from '../../baseComponents/form/formcontrol2/OFFormControlInputFormiked'
+import OFFormControlInputFormiked from '../../baseComponents/form/formControl/OFFormControlInputFormiked'
 import OFButton from '../../baseComponents/button/OFButton'
 import Box from '@material-ui/core/Box'
 import { FormikObserver } from '../../baseComponents/form/formik/FormikObserver'
