+++ src/admin/project/talks/TalkAddEditPanel.js
@@ -3,12 +3,12 @@ import { Field, Form, Formik } from 'formik'
 import { object, string, array } from 'yup'
 import SidePanelLayout from '../../baseComponents/layouts/SidePanelLayout'
 import OFButton from '../../baseComponents/button/OFButton'
-import OFFormControl from '../../baseComponents/form/formcontrol2/OFFormControl'
+import OFFormControl from '../../baseComponents/form/formControl/OFFormControl'
 import { MuiPickersUtilsProvider } from '@material-ui/pickers'
 import LuxonUtils from '@date-io/luxon'
-import OFAutoComplete from '../../baseComponents/form/autocomplete2/OFAutoComplete'
-import OFFormControlInputFormiked from '../../baseComponents/form/formcontrol2/OFFormControlInputFormiked'
-import OFDateTimePicker from '../../baseComponents/form/datetimepicker2/OFDateTimePicker'
+import OFAutoComplete from '../../baseComponents/form/autoComplete/OFAutoComplete'
+import OFFormControlInputFormiked from '../../baseComponents/form/formControl/OFFormControlInputFormiked'
+import OFDateTimePicker from '../../baseComponents/form/dateTimePicker/OFDateTimePicker'
 import SpeakerAddEditPanel from '../speakers/SpeakerAddEditPanel'
 import { useTranslation } from 'react-i18next'
 import { DateTime } from 'luxon'
